(PC-17104)[API] test: move test to appropriate file
